Rename synthwave to radiant
